Split table commands into a separate module
