+++ src/schema-table/index.js
@@ -1,5 +1,5 @@
 const {Block, Attribute, Fragment, Slice} = require("../model")
-const {Step, StepResult, PosMap, ReplaceStep} = require("../transform")
+const {Step, StepResult, PosMap} = require("../transform")
 const {copyObj} = require("../util/obj")
 
 // ;; A table node type. Has one attribute, **`columns`**, which holds
@@ -260,98 +260,3 @@ class RemoveColumnStep extends Step {
 exports.RemoveColumnStep = RemoveColumnStep
 
 Step.jsonID("removeTableColumn", RemoveColumnStep)
-
-// Table-related command functions
-
-// FIXME this module doesn't depend on the editor module. Do these
-// functions which take an editor belong here? Can't express
-// moveToNextCell without access to Selection
-
-function findRow($pos, pred) {
-  for (let d = $pos.depth; d > 0; d--)
-    if ($pos.node(d).type instanceof TableRow && (!pred || pred(d))) return d
-  return -1
-}
-
-// :: (ProseMirror, ?bool) → bool
-// Command function that adds a column before the column with the
-// selection.
-function addColumnBefore(pm, apply) {
-  let $from = pm.selection.$from, cellFrom
-  let rowDepth = findRow($from, d => cellFrom = d == $from.depth ? $from.nodeBefore : $from.node(d + 1))
-  if (rowDepth == -1) return false
-  if (apply !== false)
-    pm.tr.step(AddColumnStep.create(pm.doc, $from.before(rowDepth - 1), $from.index(rowDepth),
-                                    cellFrom.type, cellFrom.attrs)).apply()
-  return true
-}
-exports.addColumnBefore = addColumnBefore
-
-// :: (ProseMirror, ?bool) → bool
-// Command function that adds a column after the column with the
-// selection.
-function addColumnAfter(pm, apply) {
-  let $from = pm.selection.$from, cellFrom
-  let rowDepth = findRow($from, d => cellFrom = d == $from.depth ? $from.nodeAfter : $from.node(d + 1))
-  if (rowDepth == -1) return false
-  if (apply !== false)
-    pm.tr.step(AddColumnStep.create(pm.doc, $from.before(rowDepth - 1),
-                                    $from.indexAfter(rowDepth) + (rowDepth == $from.depth ? 1 : 0),
-                                    cellFrom.type, cellFrom.attrs)).apply()
-  return true
-}
-exports.addColumnAfter = addColumnAfter
-
-// :: (ProseMirror, ?bool) → bool
-// Command function that removes the column with the selection.
-function removeColumn(pm, apply) {
-  let $from = pm.selection.$from
-  let rowDepth = findRow($from, d => $from.node(d).childCount > 1)
-  if (rowDepth == -1) return false
-  if (apply !== false)
-    pm.tr.step(RemoveColumnStep.create(pm.doc, $from.before(rowDepth - 1), $from.index(rowDepth))).apply()
-  return true
-}
-exports.removeColumn = removeColumn
-
-function addRow(pm, apply, side) {
-  let $from = pm.selection.$from
-  let rowDepth = findRow($from)
-  if (rowDepth == -1) return false
-  if (apply !== false) {
-    let exampleRow = $from.node(rowDepth)
-    let cells = [], pos = side < 0 ? $from.before(rowDepth) : $from.after(rowDepth)
-    exampleRow.forEach(cell => cells.push(cell.type.createAndFill(cell.attrs)))
-    let row = exampleRow.copy(Fragment.from(cells))
-    pm.tr.step(new ReplaceStep(pos, pos, new Slice(Fragment.from(row), 0, 0))).apply()
-  }
-  return true
-}
-
-// :: (ProseMirror, ?bool) → bool
-// Command function that adds a row after the row with the
-// selection.
-function addRowBefore(pm, apply) {
-  return addRow(pm, apply, -1)
-}
-exports.addRowBefore = addRowBefore
-
-// :: (ProseMirror, ?bool) → bool
-// Command function that adds a row before the row with the
-// selection.
-function addRowAfter(pm, apply) {
-  return addRow(pm, apply, 1)
-}
-exports.addRowAfter = addRowAfter
-
-// :: (ProseMirror, ?bool) → bool
-// Command function that removes the row with the selection.
-function removeRow(pm, apply) {
-  let $from = pm.selection.$from
-  let rowDepth = findRow($from, d => $from.node(d - 1).childCount > 1)
-  if (rowDepth == -1) return false
-  if (apply !== false)
-    pm.tr.step(new ReplaceStep($from.before(rowDepth), $from.after(rowDepth), Slice.empty)).apply()
-  return true
-}
-exports.removeRow = removeRow
